requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flatbuffers
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras
Keras-Preprocessing
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard
tensorflow-cpu
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor==1.1.0
typing_extensions==4.0.1
urllib3
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
chardet
pillow
scipy
PyYAML
